# use the maximum distance to hyper rectangle to add all points in the subtree for `inrange` calculations