--- a/src/hyperrectangles.jl
+++ b/src/hyperrectangles.jl
@@ -53,3 +53,23 @@ get_min_distance_no_end(m, rec, point) =
     diff_tot = eval_diff(M, split_diff_pow, ddiff_pow, split_dim)
     return old_min + diff_tot
 end
+
+# Compute per-dimension contributions for max distance
+function get_max_distance_contributions(m::Metric, rec::HyperRectangle{V}, point::AbstractVector{T}) where {V,T}
+    p = Distances.parameters(m)
+    return V(
+        @inbounds begin
+                v = distance_function_max(point[dim], rec.maxes[dim], rec.mins[dim])
+                p === nothing ? eval_op(m, v, zero(T)) : eval_op(m, v, zero(T), p[dim])
+            end for dim in eachindex(point)
+    )
+end
+
+# Compute single dimension contribution for max distance
+function get_max_distance_contribution_single(m::Metric, point_dim, min_bound::T, max_bound::T, dim::Integer) where {T}
+    v = distance_function_max(point_dim, max_bound, min_bound)
+    p = Distances.parameters(m)
+    return p === nothing ? eval_op(m, v, zero(T)) : eval_op(m, v, zero(T), p[dim])
+end
+
+