All the tests I have done shows a reduction in flops and improved performance with this. As the radius increase this becomes more and more beneficial.

For example with

```julia
using StaticArrays, NearestNeighbors, StableRNGs, BenchmarkTools

# For uniform distribution in unit hypercube [0,1]^d,
# the expected fraction of points within radius r is approximately:
# Volume of hypersphere with radius r = π^(d/2) * r^d / Γ(d/2 + 1)
# Solving for r: r = (target_fraction * Γ(d/2 + 1) / π^(d/2))^(1/d)

function compute_radius_for_fraction(dim::Int, target_fraction::Float64)
    # Manual computation of Γ(dim/2 + 1) for small integer dimensions
    if dim % 2 == 0
        # Even dimension: Γ(n+1) = n! where n = dim/2
        n = dim ÷ 2
        gamma_term = factorial(n)
    else
        # Odd dimension: Γ(n+0.5) = (2n)! * √π / (4^n * n!) where n = (dim-1)/2
        n = (dim - 1) ÷ 2
        gamma_term = factorial(2*n) * sqrt(π) / (4^n * factorial(n))
    end

    pi_term = π^(dim/2)

    r = (target_fraction * gamma_term / pi_term)^(1/dim)
    return r
end

for dim in (2,3,5)
    data = rand(StableRNG(1), dim, 10^6);
    for target_fraction = [0.001, 0.01, 0.05, 0.1, 0.2]
        r = compute_radius_for_fraction(dim, target_fraction)
        point = rand(StableRNG(2), dim);
        x = KDTree(data)
        l = @btime inrange($x, $point, $r)
    end
end
```

I get

master
```
  3.614 μs (8 allocations: 21.83 KiB)
  20.957 μs (12 allocations: 121.91 KiB)
  86.362 μs (16 allocations: 812.55 KiB)
  141.991 μs (16 allocations: 812.55 KiB)
  247.547 μs (18 allocations: 1.83 MiB)
  3.082 μs (6 allocations: 7.54 KiB)
  18.203 μs (12 allocations: 121.91 KiB)
  72.174 μs (14 allocations: 326.45 KiB)
  137.667 μs (16 allocations: 812.55 KiB)
  240.863 μs (16 allocations: 812.55 KiB)
  8.324 μs (6 allocations: 7.54 KiB)
  32.429 μs (10 allocations: 53.87 KiB)
  111.277 μs (12 allocations: 121.91 KiB)
  193.624 μs (14 allocations: 326.45 KiB)
  329.769 μs (14 allocations: 326.45 KiB)
```

PR
```
  2.792 μs (8 allocations: 21.83 KiB)
  12.316 μs (12 allocations: 121.91 KiB)
  46.090 μs (16 allocations: 812.55 KiB)
  70.028 μs (16 allocations: 812.55 KiB)
  127.682 μs (18 allocations: 1.83 MiB)
  3.096 μs (6 allocations: 7.54 KiB)
  15.067 μs (12 allocations: 121.91 KiB)
  53.540 μs (14 allocations: 326.45 KiB)
  99.493 μs (16 allocations: 812.55 KiB)
  163.672 μs (16 allocations: 812.55 KiB)
  8.093 μs (6 allocations: 7.54 KiB)
  35.132 μs (10 allocations: 53.87 KiB)
  120.850 μs (12 allocations: 121.91 KiB)
  209.334 μs (14 allocations: 326.45 KiB)
  345.637 μs (14 allocations: 326.45 KiB)
```